Review: move Nel to base_models
