+++ components/renku_data_services/search/user_query.py
@@ -8,12 +8,12 @@ from dataclasses import dataclass, field
 from dataclasses import field as data_field
 from datetime import date, datetime, time, timedelta, tzinfo
 from enum import StrEnum
-from typing import Any, Self, override
+from typing import Self, override
 
 from renku_data_services.authz.models import Role, Visibility
 from renku_data_services.base_models.core import NamespaceSlug
+from renku_data_services.base_models.nel import Nel
 from renku_data_services.namespace.models import UserNamespace
-from renku_data_services.search.nel import Nel
 from renku_data_services.solr.entity_documents import EntityType
 from renku_data_services.solr.solr_client import SortDirection
 from renku_data_services.users.models import UserInfo
