+++ test/components/renku_data_services/search/test_db.py
@@ -8,12 +8,12 @@ from ulid import ULID
 
 from renku_data_services.authz.models import Visibility
 from renku_data_services.base_models.core import APIUser, NamespacePath, NamespaceSlug, ProjectPath, ProjectSlug
+from renku_data_services.base_models.nel import Nel
 from renku_data_services.data_connectors.models import CloudStorageCore, DataConnector
 from renku_data_services.migrations.core import run_migrations_for_app
 from renku_data_services.namespace.models import ProjectNamespace, UserNamespace
 from renku_data_services.search.db import DbUsernameResolve, SearchUpdatesRepo
 from renku_data_services.search.models import DeleteDoc
-from renku_data_services.search.nel import Nel
 from renku_data_services.search.user_query import UserId, Username
 from renku_data_services.solr.entity_documents import DataConnector as DataConnectorDoc
 from renku_data_services.solr.entity_documents import User as UserDoc
