+++ test/components/renku_data_services/search/test_solr_token.py
@@ -5,7 +5,7 @@ from zoneinfo import ZoneInfo
 
 import renku_data_services.search.solr_token as st
 from renku_data_services.authz.models import Visibility
-from renku_data_services.search.nel import Nel
+from renku_data_services.base_models.nel import Nel
 from renku_data_services.solr.entity_documents import EntityType
 from renku_data_services.solr.solr_schema import FieldName
