+++ test/components/renku_data_services/search/test_solr_user_query.py
@@ -8,7 +8,7 @@ import pytest
 
 import renku_data_services.search.solr_token as st
 from renku_data_services.authz.models import Role, Visibility
-from renku_data_services.search.nel import Nel
+from renku_data_services.base_models.nel import Nel
 from renku_data_services.search.solr_user_query import AuthAccess, Context, SolrUserQuery, _LuceneQueryTransform
 from renku_data_services.search.solr_user_query import LuceneQueryInterpreter as L
 from renku_data_services.search.user_query import (
